DGL_N2G/requirements.txt
dgl==0.9.1
dgl_cuda11.3==0.8.1
numpy==1.19.2
PyYAML==6.0
torch==1.10.2
